+++ lib/representable/hash.rb
@@ -9,18 +9,13 @@ module Representable
     def self.included(base)
       base.class_eval do
         include Representable # either in Hero or HeroRepresentation.
-        extend ClassMethods # DISCUSS: do that only for classes?
+        extend ClassMethods
         register_feature Representable::Hash
       end
     end
 
 
     module ClassMethods
-      def from_hash(*args, &block)
-        warn "[Representable] Deprecation warning: Class methods ::from_hash, ::from_json, ::from_xml, etc. will be removed in 2.0. Please use Model.new.from_json instead. Keep smiling."
-        create_represented(*args, &block).from_hash(*args)
-      end
-
       def collection_representer_class
         Collection
       end